Fix wrong order of coordinates
To access a specific cell, it has to be addressed like field[y][x]. Note
that the order of the coordinates if different from the usual way.

This was not correctly done in our implementations. To prevent this from
happening in the future, we should use functions that capsulate this
behavior in as much places as possible.